% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_agencies.R
\docType{data}
\name{fe_agencies}
\alias{fe_agencies}
\title{Agency arguments for Federal Reporters}
\format{A character vector}
\usage{
fe_agencies
}
\description{
Agency arguments for Federal Reporters
}
\note{
See \url{https://api.federalreporter.nih.gov/html/Agencies.html}
}
\keyword{datasets}
