% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_base_url.R
\name{fe_base_url}
\alias{fe_base_url}
\title{Federal Exporter Base URL}
\usage{
fe_base_url(secure = TRUE)
}
\arguments{
\item{secure}{Should https be used (may be necessary)}
}
\value{
Character vector (length 1)of URL
}
\description{
Federal Exporter Base URL
}
\examples{
fe_base_url()
}
